+++ shared/types/colony.js
@@ -1,6 +1,0 @@
-"use strict";
-/**
- * Types for colony management - shared between web app and Firebase functions
- */
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=colony.js.map